backend/requirements.txt
fastapi
uvicorn
pypdf
pdfrw
python-multipart
python-dotenv
groq
pymupdf
ollama
